Modernize the build process for Windows

+++ source/sdl4ogre/sdlwindowhelper.cpp
@@ -1,4 +1,4 @@
-#include "pch.h"
+
 #include "sdlwindowhelper.hpp"
 
 #include <OgreStringConverter.h>
@@ -32,7 +32,7 @@ SDLWindowHelper::SDLWindowHelper (SDL_Window* window, int w, int h,
 #ifdef WIN32
 	case SDL_SYSWM_WINDOWS:
 		// Windows code
-		winHandle = Ogre::StringConverter::toString((unsigned long)wmInfo.info.win.window);
+		winHandle = Ogre::StringConverter::toString((uintptr_t)wmInfo.info.win.window);
 		break;
 #elif __MACOSX__
 	case SDL_SYSWM_COCOA:
@@ -45,7 +45,7 @@ SDLWindowHelper::SDLWindowHelper (SDL_Window* window, int w, int h,
 #else
 	case SDL_SYSWM_X11:
         param = "parentWindowHandle";
-		winHandle = Ogre::StringConverter::toString((unsigned long)wmInfo.info.x11.window);
+		winHandle = Ogre::StringConverter::toString((uintptr_t)wmInfo.info.x11.window);
 		break;
 #endif
 	default:
